removed util package and small refactoring
