+++ src/main/resources/bigtable-autoscaler.conf
@@ -1,24 +1,24 @@
 # Configuration for autoscaler
 
 http.server {
-    port: 8080
-    port: ${?PORT}
+  port: 8080
+  port: ${?PORT}
 }
 
 ffwd {
-    host: "127.0.0.1"
-    port: 19091
+  host: "127.0.0.1"
+  port: 19091
 }
 
 database {
-    username: ${DB_USERNAME}
-    password: ${DB_PASSWORD}
-    jdbcUrl: ${JDBC_URL}
-    maxConnectionPool: 16
+  username: ${DB_USERNAME}
+  password: ${DB_PASSWORD}
+  jdbcUrl: ${JDBC_URL}
+  maxConnectionPool: 16
 }
 
 additionalPackages = []
 
-additionalClasses = "com.spotify.autoscaler.util.ThrowableMapper"
+additionalClasses = "com.spotify.autoscaler.ThrowableMapper"
 
 clusterFilter = ""
