+++ src/test/java/com/spotify/autoscaler/AutoscalerTest.java
@@ -37,10 +37,10 @@ import com.spotify.autoscaler.client.StackdriverClient;
 import com.spotify.autoscaler.db.BigtableCluster;
 import com.spotify.autoscaler.db.BigtableClusterBuilder;
 import com.spotify.autoscaler.db.Database;
+import com.spotify.autoscaler.db.ErrorCode;
 import com.spotify.autoscaler.filters.AllowAllClusterFilter;
 import com.spotify.autoscaler.filters.ClusterFilter;
 import com.spotify.autoscaler.metric.AutoscalerMetrics;
-import com.spotify.autoscaler.util.ErrorCode;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.Optional;
