+++ src/test/java/com/spotify/autoscaler/metric/AutoscalerMetricsTest.java
@@ -28,8 +28,8 @@ import com.codahale.metrics.Gauge;
 import com.google.common.collect.ImmutableList;
 import com.spotify.autoscaler.db.BigtableCluster;
 import com.spotify.autoscaler.db.BigtableClusterBuilder;
+import com.spotify.autoscaler.db.ErrorCode;
 import com.spotify.autoscaler.db.PostgresDatabase;
-import com.spotify.autoscaler.util.ErrorCode;
 import com.spotify.metrics.core.MetricId;
 import com.spotify.metrics.core.SemanticMetricRegistry;
 import java.util.Collections;
