+++ src/test/resources/test-service.conf
@@ -1,1 +1,1 @@
-additionalClasses = "com.spotify.autoscaler.util.ThrowableMapper"
+additionalClasses = "com.spotify.autoscaler.ThrowableMapper"
